+++ lib/src/util/_pieces.dart
@@ -1,0 +1,11 @@
+class Pieces implements Sink<List<int>> {
+  Pieces(this._held);
+
+  final List<List<int>> _held;
+
+  @override
+  void add(List<int> data) => _held.add(data);
+
+  @override
+  void close() {}
+}